refactor: 이미지 형식을 .png에서 .svg로 변경
